clean up wpas.php, refactor into new files
